+++ libhesai/UdpParserGpu/src/udp2_6_parser_gpu.cu
@@ -342,7 +342,7 @@ int Udp2_6ParserGpu<T_Point>::LoadCorrectionFile(std::string lidar_correction_fi
     LogDebug("Open correction file success");
     int length = 0;
     fin.seekg(0, std::ios::end);
-    length = fin.tellg();
+    length = static_cast<int>(fin.tellg());
     fin.seekg(0, std::ios::beg);
     char *buffer = new char[length];
     fin.read(buffer, length);
