+++ libhesai/UdpProtocol/udp_protocol_v2_4.h
@@ -192,7 +192,7 @@ namespace hesai
      
       uint8_t GetLaserNum() const { return m_u8LaserNum; }
       uint8_t GetBlockNum() const { return m_u8BlockNum; }
-      double GetDistUnit() const { return m_u8DistUnit / 1000.f; }
+      float GetDistUnit() const { return m_u8DistUnit / 1000.f; }
       uint8_t GetEchoCount() const { return m_u8EchoCount; }
       uint8_t GetSeqNum() const { return m_u8SeqNum; }
